backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
alembicaiofiles @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_0ar7qj9rcj/croot/aiofiles_1683773588297/work
Pillow @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_124pvldps5/croot/pillow_1695134106500/work
python-magic-bin==0.4.14
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0









